+++ static/app/views/dashboards/datasetConfig/releases.tsx
@@ -6,11 +6,9 @@ import {doSessionsRequest} from 'sentry/actionCreators/sessions';
 import type {Client} from 'sentry/api';
 import {t} from 'sentry/locale';
 import type {PageFilters, SelectValue} from 'sentry/types/core';
-import type {Series} from 'sentry/types/echarts';
 import type {Organization, SessionApiResponse} from 'sentry/types/organization';
 import type {SessionsMeta} from 'sentry/types/sessions';
 import {SessionField} from 'sentry/types/sessions';
-import {defined} from 'sentry/utils';
 import type {TableData} from 'sentry/utils/discover/discoverQuery';
 import {getFieldRenderer} from 'sentry/utils/discover/fieldRenderers';
 import type {
@@ -26,7 +24,7 @@ import {FieldValueKind} from 'sentry/views/discover/table/types';
 import type {Widget, WidgetQuery} from '../types';
 import {DisplayType} from '../types';
 import {getWidgetInterval} from '../utils';
-import {getSeriesName} from '../utils/transformSessionsResponseToSeries';
+import {transformSessionsResponseToSeries} from '../utils/transformSessionsResponseToSeries';
 import {
   changeObjectValuesToTypes,
   getDerivedMetrics,
@@ -34,7 +32,6 @@ import {
 } from '../utils/transformSessionsResponseToTable';
 import {ReleaseSearchBar} from '../widgetBuilder/buildSteps/filterResultsStep/releaseSearchBar';
 import {
-  DERIVED_STATUS_METRICS_PATTERN,
   DerivedStatusFields,
   DISABLED_SORT,
   FIELD_TO_METRICS_EXPRESSION,
@@ -44,7 +41,6 @@ import {
   TAG_SORT_DENY_LIST,
 } from '../widgetBuilder/releaseWidget/fields';
 import {
-  derivedMetricsToField,
   requiresCustomReleaseSorting,
   resolveDerivedStatusFields,
 } from '../widgetCard/releaseWidgetQueries';
@@ -305,85 +301,6 @@ export function transformSessionsResponseToTable(
     fields: changeObjectValuesToTypes(omit(singleRow, 'id')),
   };
   return {meta, data: rows};
-}
-
-export function transformSessionsResponseToSeries(
-  data: SessionApiResponse,
-  widgetQuery: WidgetQuery
-) {
-  if (data === null) {
-    return [];
-  }
-
-  const queryAlias = widgetQuery.name;
-
-  const useSessionAPI = widgetQuery.columns.includes('session.status');
-  const {derivedStatusFields: requestedStatusMetrics, injectedFields} =
-    resolveDerivedStatusFields(
-      widgetQuery.aggregates,
-      widgetQuery.orderby,
-      useSessionAPI
-    );
-
-  const results: Series[] = [];
-
-  if (!data.groups.length) {
-    return [
-      {
-        seriesName: `(${t('no results')})`,
-        data: data.intervals.map(interval => ({
-          name: interval,
-          value: 0,
-        })),
-      },
-    ];
-  }
-
-  data.groups.forEach(group => {
-    Object.keys(group.series).forEach(field => {
-      // if `sum(session)` or `count_unique(user)` are not
-      // requested as a part of the payload for
-      // derived status metrics through the Sessions API,
-      // they are injected into the payload and need to be
-      // stripped.
-      if (!injectedFields.includes(derivedMetricsToField(field))) {
-        results.push({
-          seriesName: getSeriesName(field, group, queryAlias),
-          data: data.intervals.map((interval, index) => ({
-            name: interval,
-            value: group.series[field]?.[index] ?? 0,
-          })),
-        });
-      }
-    });
-    // if session.status is a groupby, some post processing
-    // is needed to calculate the status derived metrics
-    // from grouped results of `sum(session)` or `count_unique(user)`
-    if (requestedStatusMetrics.length && defined(group.by['session.status'])) {
-      requestedStatusMetrics.forEach(status => {
-        const result = status.match(DERIVED_STATUS_METRICS_PATTERN);
-        if (result) {
-          let metricField: string | undefined = undefined;
-          if (group.by['session.status'] === result[1]) {
-            if (result[2] === 'session') {
-              metricField = 'sum(session)';
-            } else if (result[2] === 'user') {
-              metricField = 'count_unique(user)';
-            }
-          }
-          results.push({
-            seriesName: getSeriesName(status, group, queryAlias),
-            data: data.intervals.map((interval, index) => ({
-              name: interval,
-              value: metricField ? group.series[metricField]?.[index] ?? 0 : 0,
-            })),
-          });
-        }
-      });
-    }
-  });
-
-  return results;
 }
 
 function fieldsToDerivedMetrics(field: string): string {
